コード整理
